% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cb_dark_blue}
\alias{cb_dark_blue}
\title{Function to generate text string of individual colors}
\usage{
cb_dark_blue()
}
\value{
This function returns the hex code for cb_dark_blue
}
\description{
Function to generate text string of individual colors
}
