% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cb_gray}
\alias{cb_gray}
\title{Function to generate text string of individual colors}
\usage{
cb_gray()
}
\value{
This function returns the hex code for cb_gray
}
\description{
Function to generate text string of individual colors
}
