% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cb_green}
\alias{cb_green}
\title{Function to generate text string of individual colors}
\usage{
cb_green()
}
\value{
This function returns the hex code for cb_green
}
\description{
Function to generate text string of individual colors
}
