% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cb_pink}
\alias{cb_pink}
\title{Function to generate text string of individual colors}
\usage{
cb_pink()
}
\value{
This function returns the hex code for cb_pink
}
\description{
Function to generate text string of individual colors
}
