% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cb_yellow}
\alias{cb_yellow}
\title{Function to generate text string of individual colors}
\usage{
cb_yellow()
}
\value{
This function returns the hex code for cb_yellow
}
\description{
Function to generate text string of individual colors
}
