% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_compassred.R
\docType{data}
\name{compassred_colors}
\alias{compassred_colors}
\title{Add CompassRed palettes to ggplot charts}
\format{An object of class \code{character} of length 13.}
\usage{
compassred_colors
}
\description{
This function allows you to add the CompassRed palette to your ggplot graphics.
}
\keyword{CR_cols}
\keyword{CR_palettes}
