% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{CR_blue}
\alias{CR_blue}
\title{Function to generate text string of individual colors}
\usage{
CR_blue()
}
\value{
This function returns the hex code for CR_blue
}
\description{
Function to generate text string of individual colors
}
