% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_compassred.R
\name{CR_cols}
\alias{CR_cols}
\title{Function to extract compassred_colors colors as hex codes}
\usage{
CR_cols(...)
}
\arguments{
\item{...}{Names of compassred_colors}
}
\description{
Function to extract compassred_colors colors as hex codes
}
