% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{CR_green}
\alias{CR_green}
\title{Function to generate text string of individual colors}
\usage{
CR_green()
}
\value{
This function returns the hex code for CR_green
}
\description{
Function to generate text string of individual colors
}
