% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{CR_red}
\alias{CR_red}
\title{Function to generate text string of individual colors}
\usage{
CR_red()
}
\value{
This function returns the hex code for CR_red
}
\description{
Function to generate text string of individual colors
}
