% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{dark_gray}
\alias{dark_gray}
\title{Function to generate text string of individual colors}
\usage{
dark_gray()
}
\value{
This function returns the hex code for dark_gray
}
\description{
Function to generate text string of individual colors
}
