% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{light_gray}
\alias{light_gray}
\title{Function to generate text string of individual colors}
\usage{
light_gray()
}
\value{
This function returns the hex code for light_gray
}
\description{
Function to generate text string of individual colors
}
