% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_compassred.R
\name{theme_compassred}
\alias{theme_compassred}
\title{Add CompassRed theme to ggplot chart}
\usage{
theme_compassred()
}
\description{
This function allows you to add the CompassRed theme to your ggplot graphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_compassred()
}
\keyword{theme_compassred}
