% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR_cols.R
\name{CR_cols}
\alias{CR_cols}
\title{Function to extract CompassRed colors as hex codes}
\usage{
CR_cols(...)
}
\arguments{
\item{...}{Names of CompassRed colors as strings}
}
\value{
Either one or many hex codes as strings
}
\description{
Function to extract CompassRed colors as hex codes
}
